RecentHTMLSnippet:Refactor the sqlite interaction to a class anyone in the program can use
