+++ WebPad/MainWindow.xaml.cs
@@ -170,7 +170,6 @@ namespace WebPad
         {
             InitializeComponent();
             this.Model = new MainWindowModel();
-            setupFileDatabase();
 
             populateListOfRecentFiles().ContinueWith((t) =>
             {
@@ -206,35 +205,8 @@ namespace WebPad
         }
 
 
-        private void setupFileDatabase()
-        {
-            this.db = new netstandardDbSQLiteHelper.Database(Properties.Settings.Default.DatabaseFilePath);
-
-            db.Command(@"
-                create table if not exists RecentFiles(
-                    FileName varchar(200) not null,
-                    FullPath varchar(2000) not null,
-                    Type varchar(100) not null
-                )
-            ");
-            db.Command(@"
-                create table if not exists HtmlSnippets(
-                    BaseFilePath varchar(2000) not null,
-                    FilePath varchar(2000) not null
-                )
-            ");
-        }
 
 
-        private File.SaveHandler.SaveType GetRecentFileTypeFromString(string text)
-        {
-            if(string.IsNullOrWhiteSpace(text))
-            {
-                return File.SaveHandler.SaveType.WebPad;
-            }
-
-            return (File.SaveHandler.SaveType)Enum.Parse(typeof(File.SaveHandler.SaveType), text);
-        }
 
 
         private Task<List<Models.RecentFileModel>> populateListOfRecentFiles()
@@ -243,17 +215,7 @@ namespace WebPad
 
             var t = new Thread(() =>
             {
-                var recentEntries = db.Query(@"
-                    select *
-                    from RecentFiles
-                ");
-
-                var files = recentEntries.Select(dict => new Models.RecentFileModel
-                {
-                    FileName = dict["FileName"] as string,
-                    Path = dict["FullPath"] as string,
-                    Type = GetRecentFileTypeFromString(dict["Type"] as string)
-                });
+                var files = Utilities.DBManager.GetAllRecentFiles();
 
                 p.SetResult(files.ToList());
             });
@@ -301,15 +263,7 @@ namespace WebPad
 
             if( !string.Equals(form.Model["FileName"] as string, file.FileName))
             {
-                db.Command(@"
-                    update RecentFiles
-                    set FileName = :name
-                    where LOWER(FullPath) = :path
-                ", new Dictionary<string, object>
-                {
-                    { ":name", form.Model["FileName"] as string },
-                    {":path", file.Path.ToLower() }
-                });
+                Utilities.DBManager.RenameRecentFile(file, form.Model["FileName"] as string);
                 file.FileName = form.Model["FileName"] as string;
             }
 
@@ -332,29 +286,10 @@ namespace WebPad
             var t = new Thread(() =>
             {
 
-                var existingEntryResult = db.Query(@"
-                    select *
-                    from RecentFiles
-                    where LOWER(FullPath) = :path
-                ", new Dictionary<string, object>
+                if(Utilities.DBManager.AddRecentFileIfNotDuplicate(file))
                 {
-                    {":path", file.Path.ToLower() }
-                });
-
-                if(!existingEntryResult.Any())
-                {
-                    db.Command(@"
-                    insert into RecentFiles(FileName, FullPath, Type)
-                    values(:name, :path, :type)
-                    ", new Dictionary<string, object>
-                    {
-                        {":name", file.FileName },
-                        {":path", file.Path },
-                        {":type", file.Type.ToString() }
-                    });
                     p.SetResult(1);
-                }
-                else
+                }else
                 {
                     p.SetResult(-1);
                 }
